move parquet-specifi logic to new parquet module and add test_parquet.py
